backend/pyproject.toml
[project]
name = "tapmenu-backend"
version = "0.1.0"
description = "QR Restaurant Platform - Backend API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "django-filter>=24.0",
    "pillow>=10.0",
    "qrcode>=7.4",
    "python-decouple>=3.8",
    "drf-spectacular>=0.27",
]
